pyproject.toml
[tool.poetry]
name = "bookstore"
version = "0.1.0"
description = "Bookstore API"
authors = ["Nicholas Bau <57227204+ncholasb@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.10"
pytest = "^6.2.5"
factory-boy = "^3.2.1"
django-rest-framework = "^0.1.0"
django-extensions = "^3.1.5"
django-debug-toolbar = "^3.2.4"

[tool.poetry.dev-dependencies]
black = {version = "^22.1.0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
